Removed rest endpoints
