+++ bin/src/server/http/endpoints.rs
@@ -6,7 +6,58 @@ use serde_json::value::Value as JsonValue;
 use serde_json;
 use serde::ser::Serialize;
 use std::u16;
+use regex;
+use std::path::Path;
+use statics;
+use std::fs::File;
+use script;
+use std::io::Read;
 use super::util::*;
+
+lazy_static! {
+    static ref SCRIPT_NAME_VALIDATOR: regex::Regex = regex::Regex::new(r"^[\w-_]+(\.lua)?$").unwrap();
+}
+
+pub fn script(req: &mut Request) -> IronResult<Response> {
+    let name: String = get_url_param(req, "name")?;
+
+    if !SCRIPT_NAME_VALIDATOR.is_match(&name[..]) {
+        return Err(create_iron_error(
+            status::BadRequest,
+            "Invalid script name".to_string(),
+        ));
+    }
+
+    let payload = read_optional_json(&mut req.body)?.unwrap_or_else(|| JsonValue::Null);
+    let path = Path::new(&statics::SCRIPT_ROOT[..]).join(name);
+
+    let mut file = File::open(&path).map_err(|_| {
+        create_iron_error(
+            status::NotFound,
+            "Could not load script".to_string(),
+        )
+    })?;
+
+    let mut contents = String::new();
+
+    file.read_to_string(&mut contents).map_err(|_| {
+        create_iron_error(
+            status::NotFound,
+            "Could not read script".to_string()
+        )
+    })?;
+
+    let value = script::run(&contents, &path, payload).map_err(|err| {
+        let error_message = format!("Script failed: {:?}", err);
+        create_iron_error(
+            status::InternalServerError,
+            error_message,
+        )
+    })?;
+
+    Ok(to_response(status::Ok, &value))
+}
+
 
 pub fn transaction(req: &mut Request) -> IronResult<Response> {
     let trans = get_transaction()?;
