+++ bin/src/server/http/mod.rs
@@ -1,6 +1,5 @@
 mod middleware;
-mod rest;
-mod transaction;
+mod endpoints;
 mod util;
 
 use iron::prelude::*;
@@ -11,21 +10,8 @@ use std::u16;
 pub fn start(port: u16) {
     let mut router = Router::new();
 
-    router.post("/transaction", transaction::transaction, "transaction");
-
-    router.put(
-        "/edge/:outbound_id/:t/:inbound_id",
-        rest::create_edge,
-        "create_edge",
-    );
-    router.get("/edge", rest::get_edges, "get_edges");
-    router.delete("/edge", rest::delete_edges, "delete_edges");
-
-    router.get("/vertex", rest::get_vertices, "get_vertices");
-    router.post("/vertex", rest::create_vertex, "create_vertex");
-    router.delete("/vertex", rest::delete_vertices, "delete_vertices");
-
-    router.post("/script/:name", rest::script, "script");
+    router.post("/script/:name", endpoints::script, "script");
+    router.post("/transaction", endpoints::transaction, "transaction");
 
     let binding = format!("0.0.0.0:{}", port);
     println!("Listening on {}", binding);
