--- bin/tests/common/mod.rs
@@ -1,5 +1,0 @@
-mod datastore;
-mod http;
-
-pub use self::datastore::*;
-pub use self::http::*;
